+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/TrackyHubApplication.java
@@ -3,11 +3,17 @@ package kernel360trackybe.trackyhub;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @EntityScan(basePackages = {
-	"kernel360.trackycore.core.common",
+	"kernel360.trackycore.core.domain.entity",
 })
+@EnableJpaRepositories(basePackages = {
+	"kernel360.trackycore.core.infrastructure",
+})
+@ComponentScan(basePackages = "kernel360")
 public class TrackyHubApplication {
 
 	public static void main(String[] args) {
