+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/application/service/CarInfoProducerService.java
@@ -1,13 +1,18 @@
 package kernel360trackybe.trackyhub.application.service;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import kernel360.trackycore.core.domain.entity.CarEntity;
+import kernel360.trackycore.core.domain.provider.CarProvider;
 import kernel360trackybe.trackyhub.application.dto.request.CarOnOffRequest;
 import kernel360trackybe.trackyhub.application.dto.request.CycleInfoRequest;
 import kernel360trackybe.trackyhub.application.dto.request.GpsHistoryMessage;
+import kernel360trackybe.trackyhub.application.dto.response.MdnBizResponse;
 import kernel360trackybe.trackyhub.config.RabbitMQProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +24,7 @@ public class CarInfoProducerService {
 
 	private final RabbitTemplate rabbitTemplate;
 	private final RabbitMQProperties rabbitMQProperties;
+	private final CarProvider carProvider;
 
 	public void sendCarStart(CarOnOffRequest carOnOffRequest) {
 
@@ -58,5 +64,13 @@ public class CarInfoProducerService {
 
 	public String getToken() {
 		return UUID.randomUUID().toString();
+	}
+
+	@Transactional(readOnly = true)
+	public List<MdnBizResponse> getMdnAndBizId() {
+		List<CarEntity> cars = carProvider.findAll();
+		return cars.stream()
+			.map(car -> MdnBizResponse.of(car.getMdn(), carProvider.findByMdn(car.getMdn()).getBiz().getId()))
+			.toList();
 	}
 }
